src/components/data_download/runtime-requirements.txt
# Generated by KFP.
kfp-kubernetes
pathlib
requests